test(holochain): reinstate gossip integration tests
